Low: stonith-helper: Skip standby_wait when target node is online.
This feature shortens the failover time in certain case. For example, this feature works if the stop operation fails.
The reason this works is that, if the target host is online, that is certain that the state of the target host is "UNCLEAN(online)". So, there is no need to wait.
If split-brain occure(IC-LAN is disconnected), It is judged that the node is not online. So, the feature does not work.